Final Set Up

+++ app/src/main/java/com/ci6222/fooddelivery/datas/vos/FoodItemVO.kt
@@ -6,10 +6,10 @@ import com.google.firebase.firestore.IgnoreExtraProperties
 class FoodItemVO(
         var food_name: String?= "",
         var food_description: String? = "",
-        var food_price: Long = 0,
+        var food_price: Double = 0.0,
         var food_rating: String? = "",
         var food_image: String? = "",
         var popular: Boolean = false,
         var itemCount: Long = 1,
-        var totalAmount: Long = 0
+        var totalAmount: Double = 0.0
 )
