+++ app/src/main/java/com/ci6222/fooddelivery/mvp/view/CheckOutView.kt
@@ -4,6 +4,6 @@ import com.ci6222.fooddelivery.datas.vos.FoodItemVO
 
 interface CheckOutView : BaseView {
     fun showOrderList(orderList: List<FoodItemVO>)
-    fun showCalculationCharge(totalAmount: Long)
+    fun showCalculationCharge(totalAmount: Double)
 
 }
